chip filter 기능 추가
